Made tick counter display (ticks / 10).

+++ GardenForce/Assets/Scripts/Map.cs
@@ -42,7 +42,7 @@ public class Map : MonoBehaviour
 
     public readonly int width = 40;
     public readonly int height = 20;
-    public float tickLimit;
+    public int tickLimit;
     public float tileSize { get; private set; }
     public float tileZ;
     public float flowerZ;
